Bug 1809998 - Part 3: Add an OnboardingView for displaying the list of onboarding cards
