+++ fenix/app/src/main/java/org/mozilla/fenix/home/sessioncontrol/SessionControlView.kt
@@ -23,12 +23,8 @@ import org.mozilla.fenix.home.Mode
 import org.mozilla.fenix.home.recentbookmarks.RecentBookmark
 import org.mozilla.fenix.home.recentvisits.RecentlyVisitedItem
 import org.mozilla.fenix.messaging.FenixMessageSurfaceId
-import org.mozilla.fenix.nimbus.OnboardingPanel
 import org.mozilla.fenix.onboarding.HomeCFRPresenter
-import org.mozilla.fenix.onboarding.OnboardingState
-import org.mozilla.fenix.onboarding.view.OnboardingAdapterItem
 import org.mozilla.fenix.utils.Settings
-import org.mozilla.fenix.nimbus.Onboarding as OnboardingConfig
 
 // This method got a little complex with the addition of the tab tray feature flag
 // When we remove the tabs from the home screen this will get much simpler again.
@@ -129,35 +125,6 @@ private fun showCollections(
 }
 
 private fun privateModeAdapterItems() = listOf(AdapterItem.PrivateBrowsingDescription)
-
-@Suppress("UnusedPrivateMember")
-private fun onboardingAdapterItems(
-    onboardingState: OnboardingState,
-    onboardingConfig: OnboardingConfig,
-): List<OnboardingAdapterItem> {
-    val items: MutableList<OnboardingAdapterItem> = mutableListOf(OnboardingAdapterItem.OnboardingHeader)
-
-    onboardingConfig.order.forEach {
-        when (it) {
-            OnboardingPanel.THEMES -> items.add(OnboardingAdapterItem.OnboardingThemePicker)
-            OnboardingPanel.TOOLBAR_PLACEMENT -> items.add(OnboardingAdapterItem.OnboardingToolbarPositionPicker)
-            // Customize FxA items based on where we are with the account state:
-            OnboardingPanel.SYNC -> if (onboardingState == OnboardingState.SignedOutNoAutoSignIn) {
-                items.add(OnboardingAdapterItem.OnboardingManualSignIn)
-            }
-            OnboardingPanel.TCP -> items.add(OnboardingAdapterItem.OnboardingTrackingProtection)
-            OnboardingPanel.PRIVACY_NOTICE -> items.add(OnboardingAdapterItem.OnboardingPrivacyNotice)
-        }
-    }
-    items.addAll(
-        listOf(
-            OnboardingAdapterItem.OnboardingFinish,
-            OnboardingAdapterItem.BottomSpacer,
-        ),
-    )
-
-    return items
-}
 
 private fun AppState.toAdapterList(settings: Settings): List<AdapterItem> = when (mode) {
     is Mode.Normal -> normalModeAdapterItems(
